// src/lib/workflow/types.ts
export type EdgeStatus = "active" | "waiting" | "skipped";

export interface RuntimeEdgeItemType {
  source: string;
  target: string;
  sourceHandle?: string;
  targetHandle?: string;
  status?: EdgeStatus;
}

export interface RuntimeNodeItemType {
  nodeId: string;
  name: string;
  intro: string;
  flowNodeType: FlowNodeTypeEnum;
  showStatus?: boolean;
  avatar?: string;
  version?: string;
  position: { x: number; y: number };
  inputs: NodeInputItemType[];
  outputs: NodeOutputItemType[];
  catchError?: boolean;
}

export interface NodeInputItemType {
  key: string;
  value?: any;
  valueType?: WorkflowIOValueTypeEnum;
  valueDesc?: string;
  label?: string;
  description?: string;
  type?: FlowNodeInputTypeEnum;
  renderTypeList?: FlowNodeInputTypeEnum[];
  list?: Array<{ label: string; value: string }>;
  required?: boolean;
  selectedTypeIndex?: number;
  connected?: boolean;
  showTargetInApp?: boolean;
  showTargetInPlugin?: boolean;
  placeholder?: string;
  maxLength?: number;
  defaultValue?: any;
  min?: number;
  max?: number;
  dynamicParamDefaultValue?: any;
  canEdit?: boolean;
  editField?: Record<string, any>;
  customInputConfig?: Record<string, any>;
  md?: string;
  mist?: boolean;
}

export interface NodeOutputItemType {
  key: string;
  label?: string;
  description?: string;
  valueType?: WorkflowIOValueTypeEnum;
  type?: string;
  list?: Array<{ label: string; value: string }>;
  targets?: Array<{ moduleId: string; key: string }>;
  defaultValue?: any;
  required?: boolean;
  value?: any;
}